1-sem/RAS/Trabalho/Parte3/PictuRAS-main/backend/tools/autocrop/requirements.txt
pika==1.3.2
python_magic==0.4.27
rembg==2.0.61
tools==0.1.9
onnxruntime==1.20.1
